粉丝用户管理界面布局

+++ src/main/java/com/jeekhan/wx/controller/FansAction.java
@@ -1,7 +1,9 @@
 package com.jeekhan.wx.controller;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import javax.validation.Valid;
 
@@ -16,13 +18,14 @@ import org.springframework.validation.ObjectError;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.RestController;
 
 import com.jeekhan.wx.api.UserMgrHandle;
 import com.jeekhan.wx.dto.Operator;
+import com.jeekhan.wx.model.FansBasic;
 import com.jeekhan.wx.model.FansTag;
 import com.jeekhan.wx.service.FansBasicService;
 import com.jeekhan.wx.service.FansTagService;
+import com.jeekhan.wx.utils.PageCond;
 /**
  * 微信用户管理服务控制类
  * 供业务服务器进行调用
@@ -61,7 +64,7 @@ public class FansAction {
 	 * @return { "errcode:0,errmsg:"ok","tag":{ "id":134,"name":"广东"} }
 	 * @throws JSONException 
 	 */
-	@RequestMapping("/createTag")
+	@RequestMapping("/tag/create")
 	@ResponseBody
 	public String createTag(@Valid FansTag fansTag,BindingResult result,Operator operator) throws JSONException {
 		JSONObject jsonRet = new JSONObject();
@@ -79,7 +82,7 @@ public class FansAction {
 				return jsonRet.toString();
 			}
 			jsonRet = UserMgrHandle.addTag(fansTag.getTagName());
-			if(!jsonRet.has("errcode")) {
+			if(jsonRet.has("tag")) {
 				jsonRet.put("errcode", 0);
 				jsonRet.put("errmsg", "ok");
 				fansTag.setTagId(jsonRet.getJSONObject("tag").getInt("id"));
@@ -98,7 +101,7 @@ public class FansAction {
 	 * @return {"errcode:0,errmsg:"ok","tags":[{"id":1, "name":"每天一罐可乐星人","count":0},{"id":2,"name":"星标组","count":0 },] } 
 	 * @throws JSONException
 	 */
-	@RequestMapping("/getAllTags")
+	@RequestMapping("/tag/getAll")
 	@ResponseBody
 	public String getAllTags() throws JSONException {
 		JSONObject jsonRet = new JSONObject();
@@ -123,7 +126,7 @@ public class FansAction {
 	 * @return {"errcode":0,"errmsg":"ok" }
 	 * @throws JSONException 
 	 */
-	@RequestMapping("/updateTag")
+	@RequestMapping("/tag/update")
 	@ResponseBody
 	public String updateTag(@Valid FansTag fansTag,BindingResult result) throws JSONException {
 		JSONObject jsonRet = new JSONObject();
@@ -146,9 +149,7 @@ public class FansAction {
 				return jsonRet.toString();
 			}
 			jsonRet = UserMgrHandle.updateTag(fansTag.getTagName(), fansTag.getTagId());
-			if(!jsonRet.has("errcode")) {
-				jsonRet.put("errcode", 0);
-				jsonRet.put("errmsg", "ok");
+			if(jsonRet.has("errcode") && jsonRet.getInt("errcode") == 0) {
 				this.fansTagService.update(fansTag);
 			}
 		} catch (Exception e) {
@@ -166,7 +167,7 @@ public class FansAction {
 	 * @return {"errcode":0, "errmsg":"ok"}
 	 * @throws JSONException
 	 */
-	@RequestMapping("/deleteTag")
+	@RequestMapping("/tag/delete")
 	@ResponseBody
 	public String deleteTag(Integer tagId) throws JSONException {
 		JSONObject jsonRet = new JSONObject();
@@ -177,9 +178,7 @@ public class FansAction {
 		}
 		try {
 			jsonRet = UserMgrHandle.deleteTag(tagId);
-			if(!jsonRet.has("errcode")) {
-				jsonRet.put("errcode", 0);
-				jsonRet.put("errmsg", "ok");
+			if(jsonRet.has("errcode") && jsonRet.getInt("errcode") == 0) {
 				this.fansTagService.delete(tagId);
 			}
 		} catch (Exception e) {
@@ -206,7 +205,7 @@ public class FansAction {
 	 *}
 	 * @throws JSONException
 	 */
-	@RequestMapping("/getUsersByTag")
+	//@RequestMapping("/getUsersByTag")
 	public JSONObject getUsersByTag(Integer tagId,@RequestParam(value="beginOpenId",required=false)String beginOpenId) throws JSONException {
 		if(tagId == null) {
 			JSONObject jsonRet = new JSONObject();
@@ -237,13 +236,14 @@ public class FansAction {
 	 * @return {  "errcode":0,   "errmsg":"ok"}
 	 * @throws JSONException
 	 */
-	@RequestMapping("/addTagToUsers")
-	public JSONObject addTagToUsers(String openIds,Integer tagId) throws JSONException {
+	@RequestMapping("/fans/addTag")
+	@ResponseBody
+	public String addTagToFanses(String openIds,Integer tagId) throws JSONException {
+		JSONObject jsonRet = new JSONObject();
 		if(openIds == null || openIds.trim().length() == 0 || tagId == null) {
-			JSONObject jsonRet = new JSONObject();
 			jsonRet.put("errcode", "-777");
 			jsonRet.put("errmsg", "用户ID列表与标签ID不可为空！");
-			return jsonRet;
+			return jsonRet.toString();
 		}
 		openIds = openIds.trim();
 		String[] arr = openIds.split(",");
@@ -252,16 +252,18 @@ public class FansAction {
 			openIdList.add(openId);
 		}
 		try {
-			JSONObject jsonRet = UserMgrHandle.addTagToUsers(openIdList, tagId);
-			return jsonRet;
+			jsonRet = UserMgrHandle.addTagToUsers(openIdList, tagId);
+			if(jsonRet.has("errcode") && jsonRet.getInt("errcode") ==0) {
+				this.fansBasicService.addTag2Fanses(openIdList, tagId);
+			}
 		} catch (Exception e) {
 			e.printStackTrace();
-			JSONObject jsonRet = new JSONObject();
 			jsonRet.put("errcode", "-666");
 			jsonRet.put("errmsg", "系统出现异常，异常信息：" + e.getMessage());
-			return jsonRet;
 		}
+		return jsonRet.toString();
 	}
+	
 	/**
 	 * 批量为用户取消标签
 	 * @param openIds	用户ID列表
@@ -269,13 +271,14 @@ public class FansAction {
 	 * @return { "errcode":0,   "errmsg":"ok"}
 	 * @throws JSONException
 	 */
-	@RequestMapping("/moveTagFromUsers")
-	public JSONObject moveTagFromUsers(String openIds,Integer tagId) throws JSONException {
+	@RequestMapping("/fans/moveTag")
+	@ResponseBody
+	public String moveTagFromFanses(String openIds,Integer tagId) throws JSONException {
+		JSONObject jsonRet = new JSONObject();
 		if(openIds == null || openIds.trim().length() == 0 || tagId == null) {
-			JSONObject jsonRet = new JSONObject();
 			jsonRet.put("errcode", "-777");
 			jsonRet.put("errmsg", "用户ID列表与标签ID不可为空！");
-			return jsonRet;
+			return jsonRet.toString();
 		}
 		openIds = openIds.trim();
 		String[] arr = openIds.split(",");
@@ -284,15 +287,16 @@ public class FansAction {
 			openIdList.add(openId);
 		}
 		try {
-			JSONObject jsonRet = UserMgrHandle.moveTagFromUsers(openIdList, tagId);
-			return jsonRet;
+			jsonRet = UserMgrHandle.moveTagFromUsers(openIdList, tagId);
+			if(jsonRet.has("errcode") && jsonRet.getInt("errcode") ==0) {
+				this.fansBasicService.removeTagFromFanses(tagId);
+			}
 		} catch (Exception e) {
 			e.printStackTrace();
-			JSONObject jsonRet = new JSONObject();
 			jsonRet.put("errcode", "-666");
 			jsonRet.put("errmsg", "系统出现异常，异常信息：" + e.getMessage());
-			return jsonRet;
 		}
+		return jsonRet.toString();
 	}
 	
 	/**
@@ -301,7 +305,7 @@ public class FansAction {
 	 * @return {"errcode":0,"errmsg":"ok"，"tagid_list":[ 134, 2] } 
 	 * @throws JSONException 
 	 */
-	@RequestMapping("/getTagsOnUser")
+	//@RequestMapping("/getTagsOnUser")
 	public JSONObject getTagsOnUser(String openId) throws JSONException {
 		if(openId == null || openId.trim().length() == 0 ) {
 			JSONObject jsonRet = new JSONObject();
@@ -333,36 +337,38 @@ public class FansAction {
 	 * @return {"errcode":0,"errmsg":"ok"}
 	 * @throws JSONException 
 	 */
-	@RequestMapping("/updateRemark4User")
-	public JSONObject updateRemark4User(String openId,String remark) throws JSONException {
+	@RequestMapping("/fans/updateRemark")
+	@ResponseBody
+	public String updateRemark4Fans(String openId,String remark) throws JSONException {
+		JSONObject jsonRet = new JSONObject();
 		if(openId == null || openId.trim().length() == 0 || remark == null || remark.trim().length() == 0) {
-			JSONObject jsonRet = new JSONObject();
 			jsonRet.put("errcode", "-777");
 			jsonRet.put("errmsg", "用户ID与备注不可为空！");
-			return jsonRet;
+			return jsonRet.toString();
 		}
 		openId = openId.trim();
 		remark = remark.trim();
 		if(remark.length() > 30) {
-			JSONObject jsonRet = new JSONObject();
 			jsonRet.put("errcode", "-888");
 			jsonRet.put("errmsg", "备注长度超过30个字符！");
-			return jsonRet;
+			return jsonRet.toString();
 		}
 		try {
-			JSONObject jsonRet = UserMgrHandle.updateRemark4User(openId, remark);
-			return jsonRet;
+			jsonRet = UserMgrHandle.updateRemark4User(openId, remark);
+			if(jsonRet.has("errcode") && jsonRet.getInt("errcode") ==0) {
+				this.fansBasicService.updateRemark(openId, remark);
+			}
 		} catch (Exception e) {
 			e.printStackTrace();
-			JSONObject jsonRet = new JSONObject();
 			jsonRet.put("errcode", "-666");
 			jsonRet.put("errmsg", "系统出现异常，异常信息：" + e.getMessage());
-			return jsonRet;
 		}
+		return jsonRet.toString();
 	}
 	
 	/**
 	 * 获取用户基本信息（包括UnionID机制）
+	 * 更新系统中的用户
 	 * @param openId	用户ID
 	 * @return
 	 * {   "errcode":0,"errmsg":"ok",
@@ -386,29 +392,30 @@ public class FansAction {
 	 * }
 	 * @throws JSONException
 	 */
-	@RequestMapping("/getUserInfo")
-	public JSONObject getUserInfo(String openId) throws JSONException {
+	@RequestMapping("/fans/getUserInfo")
+	@ResponseBody
+	public String getUserInfo(String openId) throws JSONException {
+		JSONObject jsonRet = new JSONObject();
 		if(openId == null || openId.trim().length() == 0 ) {
-			JSONObject jsonRet = new JSONObject();
 			jsonRet.put("errcode", "-777");
 			jsonRet.put("errmsg", "用户ID不可为空！");
-			return jsonRet;
+			return jsonRet.toString();
 		}
 		openId = openId.trim();
 		try {
-			JSONObject jsonRet = UserMgrHandle.getUserInfo(openId);
+			jsonRet = UserMgrHandle.getUserInfo(openId);
 			if(!jsonRet.has("errcode")) {
 				jsonRet.put("errcode", 0);
 				jsonRet.put("errmsg", "ok");
+				//更新用户信息
+				
 			}
-			return jsonRet;
 		} catch (Exception e) {
 			e.printStackTrace();
-			JSONObject jsonRet = new JSONObject();
 			jsonRet.put("errcode", "-666");
 			jsonRet.put("errmsg", "系统出现异常，异常信息：" + e.getMessage());
-			return jsonRet;
 		}
+		return jsonRet.toString();
 	}
 	
 	/**
@@ -424,7 +431,7 @@ public class FansAction {
 	 * }
 	 * @throws JSONException
 	 */
-	@RequestMapping("/getUsers")
+	//@RequestMapping("/getUsers")
 	public JSONObject getUsers(@RequestParam(value="beginOpenId",required=false)String beginOpenId) throws JSONException {
 		if(beginOpenId != null) {
 			beginOpenId = beginOpenId.trim();
@@ -460,8 +467,8 @@ public class FansAction {
 	 * }
 	 * @throws JSONException
 	 */
-	@RequestMapping("/getBlackUsers")
-	public JSONObject getBlackUsers(@RequestParam(value="beginOpenId",required=false)String beginOpenId) throws JSONException {
+	//@RequestMapping("/getBlackFans")
+	public JSONObject getBlackFans(@RequestParam(value="beginOpenId",required=false)String beginOpenId) throws JSONException {
 		beginOpenId = beginOpenId.trim();
 		try {
 			JSONObject jsonRet = UserMgrHandle.getBlackUsers(beginOpenId);
@@ -486,36 +493,37 @@ public class FansAction {
 	 * @return {"errcode":40013,"errmsg":"invalid appid"}
 	 * @throws JSONException 
 	 */
-	@RequestMapping("/blackUsers")
-	public JSONObject blackUsers(String openIds) throws JSONException {
+	@RequestMapping("/fans/black")
+	@ResponseBody
+	public String blackFanses(String openIds) throws JSONException {
+		JSONObject jsonRet = new JSONObject();
 		if(openIds == null || openIds.trim().length() == 0 ) {
-			JSONObject jsonRet = new JSONObject();
 			jsonRet.put("errcode", "-777");
 			jsonRet.put("errmsg", "用户ID列表不可为空！");
-			return jsonRet;
+			return jsonRet.toString();
 		}
 		openIds = openIds.trim();
 		String[] arr = openIds.split(",");
 		if(arr.length > 0) {
-			JSONObject jsonRet = new JSONObject();
 			jsonRet.put("errcode", "-999");
 			jsonRet.put("errmsg", "一次只能拉黑20个用户！");
-			return jsonRet;
+			return jsonRet.toString();
 		}
 		List<String> openIdList = new ArrayList<String>();
 		for(String openId:arr) {
 			openIdList.add(openId);
 		}
 		try {
-			JSONObject jsonRet = UserMgrHandle.blackUsers(openIdList);
-			return jsonRet;
+			jsonRet = UserMgrHandle.blackUsers(openIdList);
+			if(jsonRet.has("errcode") && jsonRet.getInt("errcode") ==0) {
+				this.fansBasicService.blackFans(openIdList);
+			}
 		} catch (JSONException e) {
 			e.printStackTrace();
-			JSONObject jsonRet = new JSONObject();
 			jsonRet.put("errcode", "-666");
 			jsonRet.put("errmsg", "系统出现异常，异常信息：" + e.getMessage());
-			return jsonRet;
 		}
+		return jsonRet.toString();
 	}
 	
 	/**
@@ -525,35 +533,132 @@ public class FansAction {
 	 * @return {"errcode":40013,"errmsg":"invalid appid"}
 	 * @throws JSONException 
 	 */
-	@RequestMapping("/unBlackUsers")
-	public JSONObject unBlackUsers(String openIds) throws JSONException {
+	@RequestMapping("/fans/unBlack")
+	@ResponseBody
+	public String unBlackFanses(String openIds) throws JSONException {
+		JSONObject jsonRet = new JSONObject();
 		if(openIds == null || openIds.trim().length() == 0 ) {
-			JSONObject jsonRet = new JSONObject();
 			jsonRet.put("errcode", "-777");
 			jsonRet.put("errmsg", "用户ID列表不可为空！");
-			return jsonRet;
+			return jsonRet.toString();
 		}
 		openIds = openIds.trim();
 		String[] arr = openIds.split(",");
 		if(arr.length > 0) {
-			JSONObject jsonRet = new JSONObject();
 			jsonRet.put("errcode", "-999");
 			jsonRet.put("errmsg", "一次只能拉黑20个用户！");
-			return jsonRet;
+			return jsonRet.toString();
 		}
 		List<String> openIdList = new ArrayList<String>();
 		for(String openId:arr) {
 			openIdList.add(openId);
 		}
 		try {
-			JSONObject jsonRet = UserMgrHandle.blackUsers(openIdList);
-			return jsonRet;
+			jsonRet = UserMgrHandle.blackUsers(openIdList);
+			if(jsonRet.has("errcode") && jsonRet.getInt("errcode") ==0) {
+				this.fansBasicService.unBlackFans(openIdList);
+			}
 		} catch (JSONException e) {
 			e.printStackTrace();
-			JSONObject jsonRet = new JSONObject();
 			jsonRet.put("errcode", "-666");
 			jsonRet.put("errmsg", "系统出现异常，异常信息：" + e.getMessage());
-			return jsonRet;
 		}
+		return jsonRet.toString();
 	}
+	
+	/**
+	 * 
+	 * @param jsonParams 查询条件JSON格式
+	 * @param pageCond	分页信息
+	 * @return
+	 * @throws JSONException 
+	 */
+	@RequestMapping("/fans/search")
+	@ResponseBody
+	public String searchAllFans(String jsonParams,PageCond pageCond) throws JSONException {
+		JSONObject jsonRet = new JSONObject();
+		try {
+			Map<String,Object> condParams = new HashMap<String,Object>();
+			if(jsonParams != null) {
+				JSONObject params = new JSONObject(jsonParams);
+				//条件整合
+				if(params.has("sex")) {
+					if(params.getString("sex") != null && (params.getString("sex").trim().length()>0)){
+						condParams.put("sex", params.getString("sex"));
+					}
+				}
+				if(params.has("province")) {
+					if(params.getString("province") != null && (params.getString("province").trim().length()>0)){
+						condParams.put("province", params.getString("province"));
+					}
+				}
+				if(params.has("city")) {
+					if(params.getString("city") != null && (params.getString("city").trim().length()>0)){
+						condParams.put("city", params.getString("city"));
+					}
+				}
+				if(params.has("channel")) {
+					if(params.getString("channel") != null && (params.getString("channel").trim().length()>0)){
+						condParams.put("channel", params.getString("channel"));
+					}
+				}
+				if(params.has("tagid")) {
+					if(params.getString("tagid") != null && (params.getString("tagid").trim().length()>0)){
+						condParams.put("tagid", params.getString("tagid"));
+					}
+				}
+				if(params.has("isBlack")) {
+					if(params.getString("isBlack") != null && (params.getString("isBlack").trim().length()>0)){
+						condParams.put("isBlack", params.getString("isBlack"));
+					}
+				}
+				if(params.has("openId")) {
+					if(params.getString("openId") != null && (params.getString("openId").trim().length()>0)){
+						condParams.put("openId", params.getString("openId"));
+					}
+				}
+				if(params.has("beginTime")) {
+					if(params.getString("beginTime") != null && (params.getString("beginTime").trim().length()>0)){
+						condParams.put("beginTime", params.getString("beginTime"));
+					}
+				}
+				if(params.has("endTime")) {
+					if(params.getString("endTime") != null && (params.getString("endTime").trim().length()>0)){
+						condParams.put("endTime", params.getString("endTime"));
+					}
+				}
+			}
+			if(pageCond == null) {
+				pageCond = new PageCond(0,20);
+			}
+			if(pageCond.getBegin()<0) {
+				pageCond.setBegin(0);
+			}
+			if(pageCond.getPageSize() < 2 || pageCond.getPageSize() > 100) {
+				pageCond.setPageSize(20);
+			}
+			
+			int cnt = this.fansBasicService.countAll(condParams);
+			pageCond.setCount(cnt);
+			if(cnt > 0) {
+			List<FansBasic> datas = this.fansBasicService.getAll(condParams, pageCond);
+				jsonRet.put("datas", new JSONArray(datas));
+			}
+			JSONObject jaonPage = new JSONObject();
+			jaonPage.put("pageSize", pageCond.getPageSize());
+			jaonPage.put("begin", pageCond.getBegin());
+			jaonPage.put("count", pageCond.getCount());
+			jsonRet.put("errcode", 0);
+			jsonRet.put("errmsg", "ok");
+			jsonRet.put("pageCond", jaonPage);
+		}catch(Exception e) {
+			e.printStackTrace();
+			jsonRet.put("errcode", -999);
+			jsonRet.put("errmsg","系统异常，异常信息：" + e.getMessage());
+		}
+		return jsonRet.toString();
+	}
+	
 }
+
+
